+++ apps/codeforafrica/src/components/FooterMenu/index.js
@@ -2,7 +2,7 @@ import PropTypes from "prop-types";
 import React from "react";
 
 import NavList from "@/codeforafrica/components/NavList";
-import NavListItem from "@/codeforafrica/components/NavList/NavListItem";
+import NavListItem from "@/codeforafrica/components/NavListItem";
 
 function FooterMenu({ menu }) {
   if (!menu?.length) {
